+++ jae_project/src/main/java/com/springboot/project/repository/ChatMessageRepository.java
@@ -11,9 +11,9 @@ import com.springboot.project.data.entity.ChatMessage;
 @Repository
 public interface ChatMessageRepository extends MongoRepository<ChatMessage , String>{
 
-	public ChatMessage findByRoomId(int roomId);
+	public ChatMessage findByRoomId(Long roomId);
 	
-	public Page<ChatMessage> findAllByRoomId(int roomId , Pageable pageable);
+	public Page<ChatMessage> findAllByRoomId(Long roomId , Pageable pageable);
 	
 	public ChatMessage findById(ObjectId objectId);
